Play video subsection
Plays the video only from the start time to end time indicated in the
PHP. (Later, this might be drawn from the exercise.)

Modified progress bar and time counters to reflect subsection of video.

Tags are still added with time relative to the whole video, not the the
selected subsection!